Created a website
I am starting the sudoku website. I will need to work on the CSS to display it and then afterwards, convert my python code into javascript so it can do the same thing. I used bootstrap since it is easy to start with and trying to create the board right now