Complete
I was finally able to create my sudoku visualizer. It is able to show the colors as it solves the sudoku or can do an instant solve and show which boxes were filled. I had some problems with the async but once I cleaned the code I was able to have a simple async and sleep function to do the visualizer